+++ Framework/Email.php
@@ -78,6 +78,11 @@ class Email extends Model
                     $user = $cu->getUserByEmail($address);
                     if (!$user) {
                         Configuration::getLogger()->debug('[mail] user not found', ["mail" => $mailing, "user" => $address]);
+                        continue;
+                    }
+                    if ($cu->isEmailExpired($user['email'])) {
+                        Configuration::getLogger()->debug('[mail] user mail expired', ["mail" => $mailing, "user" => $user['email']]);
+                        continue;
                     }
                     if ($user && $cm->unsubscribed($user["id"], $mailingInfo[1], $mailingInfo[0])) {
                         Configuration::getLogger()->debug('[mail] user unsubscribed', ["mail" => $mailing, "user" => $address]);
@@ -289,13 +294,18 @@ class Email extends Model
             $toAddress = $params["email"];
             $subject = CoreTranslator::AccountPendingCreationSubject($lang);
             $content = CoreTranslator::AccountPendingCreationEmail($lang, $params["jwt"], $params["url"]);
+        } elseif ($origin == "user_email_confirm") {
+            $fromName = "Platform-Manager";
+            $toAddress = $params["email"];
+            $subject = CoreTranslator::AccountEmailPendingConfirmSubject($lang);
+            $content = CoreTranslator::AccountEmailPendingConfirmEmail($lang, $params["jwt"], $params["url"]);
         } else {
             Configuration::getLogger()->error(
                 "notifyUserByEmail",
                 ["message" => "origin parameter is not set properly", "origin" => $origin]
             );
         }
-        $this->sendEmail($from, $fromName, $toAddress, $subject, $content, false);
+        $this->sendEmail($from, $fromName, $toAddress, $subject, $content, false, bcc: false);
     }
 
     /**
@@ -320,4 +330,5 @@ class Email extends Model
         $mailerSetCopyToFrom = $modelConfig->getParamSpace("MailerSetCopyToFrom", $spaceId, 0);
         return ($mailerSetCopyToFrom == 1);
     }
+
 }