+++ Framework/Events.php
@@ -46,6 +46,7 @@ require_once 'Modules/rating/Model/RatingTranslator.php';
 use PhpAmqpLib\Connection\AMQPStreamConnection;
 use PhpAmqpLib\Message\AMQPMessage;
 use Sentry\Event;
+use Firebase\JWT\JWT;
 
 class EventModel extends Model
 {
@@ -724,6 +725,36 @@ class EventHandler
         }
     }
 
+    public function userDisable($msg)
+    {
+        $this->logger->debug('[userDisable]');
+        $model = new CoreUser();
+        $model->disableUsers($msg['user_desactivate'], $msg['remove'], $msg['space']['id'], false);
+    }
+
+    public function userMailExpire()
+    {
+        $this->logger->debug('[userMailExpire]');
+        $model = new CoreUser();
+        $model->disableUnconfirmedEmails();
+    }
+
+    public function userMailCheck()
+    {
+        $this->logger->debug('[userMailCheck]');
+        $model = new CoreUser();
+        $userSettingsModel = new CoreUserSettings();
+        $users = $model->getExpiringEmails(Configuration::get('email_expire_delay', 30));
+        
+        foreach ($users as $user) {
+            $lang = Configuration::get('lang', 'en');
+            $lang = $userSettingsModel->getUserSetting($user['id'], "language", $lang);
+            Utils::requestEmailConfirmation($user['id'], $user['email'], $lang);
+            $model->newEmailCallForConfirmation($user['id']);
+        }
+    }
+
+
     /**
      * Handle message from rabbitmq
      *
@@ -803,6 +834,15 @@ class EventHandler
                 case Events::ACTION_RATING_CAMPAIGN_NEW:
                     $this->campaignRequest($data);
                     break;
+                case Events::ACTION_USER_DISABLE:
+                    $this->userDisable($data);
+                    break;
+                case Events::ACTION_USER_MAILEXPIRE:
+                    $this->userMailExpire();
+                    break;
+                case Events::ACTION_USER_MAILCHECK:
+                    $this->userMailCheck();
+                    break;
                 default:
                     $this->logger->error('[message] unknown message', ['action' => $data]);
                     $ok = false;
@@ -853,6 +893,10 @@ class Events
     public const ACTION_STATISTICS_REQUEST = 800;
 
     public const ACTION_RATING_CAMPAIGN_NEW = 900;
+
+    public const ACTION_USER_DISABLE = 1000;
+    public const ACTION_USER_MAILEXPIRE = 1001;
+    public const ACTION_USER_MAILCHECK = 1002;
 
     private static $connection;
     private static $channel;