+++ Modules/core/Model/CoreTranslator.php
@@ -158,6 +158,27 @@ class CoreTranslator
         return "Please confirm your registration at the following link: <a href=\"$confirmUrl\">$confirmUrl</a><br/>Link will expire in 2 days.";
     }
 
+
+    public static function AccountEmailPendingConfirmSubject($lang)
+    {
+        $str = "[pfm] ";
+        if ($lang == "fr") {
+            return $str . "Mail en attente de confirmation";
+        }
+        return $str . "Pending email confirmation";
+    }
+
+    public static function AccountEmailPendingConfirmEmail($lang, $jwt, $url)
+    {
+        $confirmUrl = $url."/core/account/emailconfirm?token=".$jwt;
+        if ($lang == "fr") {
+            $footer = "<p>Ceci est un mail automatique, envoyé à interval régulier, pour vérifier la validité de votre mail, merci de votre compréhension.</p>";
+            return "Merci de confirmer votre mail en allant sur le lien suivant: <a href=\"$confirmUrl\">$confirmUrl</a><br/><p>Le lien expirera dans 2 jours.</p><p>Vous pouvez également vérifier et valider votre mail dans votre compte.</p>".$footer;
+        }
+        $footer = "<p>This is an automated email sent at regular interval to check the validity of your email. Thank you for your comprehension.</p>";
+        return "Please confirm your email at the following link: <a href=\"$confirmUrl\">$confirmUrl</a><br/><p>Link will expire in 2 days</p><p>You can also check and validate your email in your account.</p>".$footer;
+    }
+
     public static function WaitingAccountMessage($lang)
     {
         if ($lang == "fr") {
@@ -2623,5 +2644,29 @@ class CoreTranslator
             return "Affichage d'éléments par tableau";
         }
         return "Displayed array elements";
+    }
+
+    public static function AccountEmailConfirmed($lang)
+    {
+        if ($lang == "fr") {
+            return "Mail confirmé";
+        }
+        return "Email confirmed";
+    }
+
+    public static function AccountEmailConfirm($lang)
+    {
+        if ($lang == "fr") {
+            return "Confirmer mails";
+        }
+        return "Confirm emails";
+    }
+
+    public static function CheckEmail($lang)
+    {
+        if ($lang == "fr") {
+            return "Confirmer mail";
+        }
+        return "Confirm email";
     }
 }