+++ db/upgrade/1670322791_add_email_validation.php
@@ -1,0 +1,15 @@
+<?php
+require_once 'Framework/Model.php';
+require_once 'Framework/Configuration.php';
+# Upgrade: add email validation
+class CoreUpgradeDB1670322791 extends Model
+{
+  public function run() {
+    Configuration::getLogger()->info("[db][upgrade] Apply add email validation");
+    $this->addColumn('core_users', 'date_email_expiration', "int", 0);
+    $this->addColumn('core_users', 'nb_email_expiration', "int", 0);
+    Configuration::getLogger()->info("[db][upgrade] Apply add email validation, done!");
+  }
+}
+$db = new CoreUpgradeDB1670322791();
+$db->run();